refactor(ui): split FieldsSchemaEditor into smaller components
